requirements.txt
anthropic>=0.40.0
flask>=3.0.0
authlib>=1.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
requests>=2.31.0
gunicorn>=21.0.0
gevent>=24.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
